+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Fixed
+- Docker, give default values for email configuration
+
 ## [1.2.0] - 2020-12-17
 
 ### Changed